BLACKOPS-332: Add proxy_to and friends to cffi backend
